+++ src/Ext/TAction/Body.cpp
@@ -10,6 +10,7 @@
 #include <Utilities/SavegameDef.h>
 
 #include <Ext/Scenario/Body.h>
+#include <Ext/Script/Body.h>
 
 //Static init
 TActionExt::ExtContainer TActionExt::ExtMap;
@@ -66,7 +67,8 @@ bool TActionExt::Execute(TActionClass* pThis, HouseClass* pHouse, ObjectClass* p
 		return TActionExt::RunSuperWeaponAtLocation(pThis, pHouse, pObject, pTrigger, location);
 	case PhobosTriggerAction::RunSuperWeaponAtWaypoint:
 		return TActionExt::RunSuperWeaponAtWaypoint(pThis, pHouse, pObject, pTrigger, location);
-
+	case PhobosTriggerAction::PrintMessageRemainingTechnos:
+		return TActionExt::PrintMessageRemainingTechnos(pThis, pHouse, pObject, pTrigger, location);
 	case PhobosTriggerAction::ToggleMCVRedeploy:
 		return TActionExt::ToggleMCVRedeploy(pThis, pHouse, pObject, pTrigger, location);
 
@@ -417,6 +419,153 @@ bool TActionExt::RunSuperWeaponAt(TActionClass* pThis, int X, int Y)
 bool TActionExt::ToggleMCVRedeploy(TActionClass* pThis, HouseClass* pHouse, ObjectClass* pObject, TriggerClass* pTrigger, CellStruct const& location)
 {
 	GameModeOptionsClass::Instance->MCVRedeploy = pThis->Param3 != 0;
+	return true;
+}
+
+bool TActionExt::PrintMessageRemainingTechnos(TActionClass* pThis, HouseClass* pHouse, ObjectClass* pObject, TriggerClass* pTrigger, CellStruct const& location)
+{
+	if (!pThis)
+		return true;
+	// Example:
+	// ID=ActionCount,[Action1],507,4,[CSFKey],[HouseIndex],[AIHousesLists Index],[AITargetTypes Index],[MesageDelay],A,[ActionX]
+	std::vector<HouseClass*> housesList;
+
+	// Obtain houses
+	int param3 = pThis->Param3;
+
+	if (param3 < 0)
+	{
+		// Pick a group of countries from [AIHousesList].
+		// Any house of the same type of the listed at [AIHousesList] will be included here
+
+		const int listIdx = pThis->Param4;
+
+		if (listIdx < 0
+		|| RulesExt::Global()->AIHousesLists.size() == 0
+		|| RulesExt::Global()->AIHousesLists.size() <= listIdx
+		|| RulesExt::Global()->AIHousesLists[listIdx].size() == 0)
+		{
+			Debug::Log("Map action %d: List [AIHousesList](%d) is empty or invalid. This action will be skipped.\n", (int)pThis->ActionKind, listIdx);
+			return true;
+		}
+
+		std::vector<HouseTypeClass*> housesTypeList;
+		housesTypeList = RulesExt::Global()->AIHousesLists.at(listIdx);
+
+		for (const auto pHouseType : housesTypeList)
+		{
+			for (auto pItem : *HouseClass::Array)
+			{
+				if (pItem->Type == pHouseType && !pItem->Defeated && !pItem->IsObserver())
+					housesList.push_back(pItem);
+			}
+		}
+
+		// Nothing to check
+		if (housesList.size() == 0)
+			return true;
+	}
+	else
+	{
+		// Check index of a single house.
+
+		if (param3 >= HouseClass::PlayerAtA && param3 <= HouseClass::PlayerAtH)
+		{
+			// Is a multiplayer house index (Player@A - Player@H) ?
+			param3 = pThis->Param3 - HouseClass::PlayerAtA;
+		}
+		else if (param3 == 8997)
+		{
+			// Is the owner of the trigger ?
+			param3 = pThis->TeamType ? pThis->TeamType->Owner->ArrayIndex : pHouse->ArrayIndex;
+		}
+		else if (param3 > 8997 || HouseClass::Array()->Count <= param3)
+		{
+			// Is a invalid index value ?
+			Debug::Log("Map action %d: Invalid house index '%d'. This action will be skipped.\n", (int)pThis->ActionKind, pThis->Param3);
+			return true;
+		}
+
+		housesList.push_back(HouseClass::Array->GetItem(param3));
+	}
+
+	// Read the ID list of technos
+	int listIdx = std::abs(pThis->Param5);
+	bool isGlobalCount = pThis->Param5 < 0;
+
+	if (RulesExt::Global()->AITargetTypesLists.size() == 0
+		|| RulesExt::Global()->AITargetTypesLists[listIdx].size() == 0
+		|| RulesExt::Global()->AITargetTypesLists.size() <= listIdx)
+	{
+		Debug::Log("Map action %d: List [AITargetTypes](%d) is empty. This action will be skipped.\n", (int)pThis->ActionKind, listIdx);
+		return true;
+	}
+
+	std::vector<TechnoTypeClass*> technosList = RulesExt::Global()->AITargetTypesLists[listIdx];
+	std::vector<int> technosRemaining;
+	int globalRemaining = 0;
+
+	// Count all valid instances
+	for (auto const pType : technosList)
+	{
+		int nRemaining = 0;
+
+		for (const auto pTechno : *TechnoClass::Array)
+		{
+			if (!ScriptExt::IsUnitAvailable(pTechno, false) || pTechno->GetTechnoType() != pType)
+				continue;
+
+			for (const auto pItem : housesList)
+			{
+				if (pTechno->Owner == pItem)
+				{
+					globalRemaining++;
+					nRemaining++;
+					break;
+				}
+			}
+		}
+
+		technosRemaining.push_back(nRemaining);
+	}
+
+	bool textToShow = false;
+	double messageDelay = pThis->Param6 <= 0 ? RulesClass::Instance->MessageDelay : pThis->Param6 / 60.0; // seconds / 60 = message delay in minutes
+	wchar_t message[2048] = { 0 };
+	wcscpy_s(message, StringTable::TryFetchString(pThis->Text, L"Remaining: "));
+
+	if (isGlobalCount)
+	{
+		if (globalRemaining > 0)
+		{
+			wchar_t strInteger[24] = { 0 };
+			swprintf_s(strInteger, L"%d", globalRemaining);
+			wcscat_s(message, strInteger);
+			textToShow = true;
+		}
+	}
+	else
+	{
+		wcscat_s(message, L"\n");
+
+		for (std::size_t i = 0; i < technosRemaining.size(); i++)
+		{
+			if (technosRemaining[i] == 0)
+				continue;
+
+			textToShow = true;
+			wcscat_s(message, technosList[i]->UIName);
+			wcscat_s(message, L": ");
+			wchar_t strInteger[24] = { 0 };
+			swprintf_s(strInteger, L"%d", technosRemaining[i]);
+			wcscat_s(message, strInteger);
+			wcscat_s(message, L"\n");
+		}
+	}
+
+	if (textToShow)
+		MessageListClass::Instance->PrintMessage(message, messageDelay, HouseClass::CurrentPlayer->ColorSchemeIndex, true);
+
 	return true;
 }